chore: more code reorganization
