+++ src/processor/processor.ts
@@ -5,12 +5,11 @@ import {
     MarkdownView,
     Notice,
     TFile,
-    editorLivePreviewField,
     type MarkdownPostProcessorContext
 } from "obsidian";
 import type { BasicRoller } from "../rollers/roller";
 import type DiceRollerPlugin from "src/main";
-import { StackRoller } from "src/rollers";
+import { StackRoller } from "src/rollers/dice/stack";
 import { isTemplateFolder } from "src/utils/util";
 import type { DiceRollerSettings } from "src/settings/settings.types";
 import { API } from "src/api/api";
