--- src/rollers/index.ts
@@ -1,5 +1,0 @@
-export * from "./dice";
-export * from "./section";
-export * from "./table";
-export * from "./line";
-export * from "./tag";
