+++ src/rollers/table/table.ts
@@ -9,8 +9,8 @@ import {
 } from "obsidian";
 
 import { TABLE_REGEX } from "src/utils/constants";
-import { StackRoller } from ".";
-import { GenericFileRoller, type ComponentLike } from "./roller";
+import { StackRoller } from "../dice/stack";
+import { GenericFileRoller } from "../roller";
 import { API } from "src/api/api";
 import type { DiceRollerSettings } from "src/settings/settings.types";
 import type { LexicalToken } from "src/lexer/lexer";
@@ -30,11 +30,7 @@ export class TableRoller extends GenericFileRoller<string> {
     lookupRanges: [range: [min: number, max: number], option: string][];
     combinedTooltip: string = "";
     prettyTooltip: string = "";
-    #components: ComponentLike[] = [];
-    override addContexts(...components: ComponentLike[]): void {
-        this.#components = components;
-        super.addContexts(...components);
-    }
+
     constructor(
         data: DiceRollerSettings,
         original: string,
@@ -164,7 +160,7 @@ export class TableRoller extends GenericFileRoller<string> {
                     continue;
                 }
                 const subRoller = maybeRoller.unwrap();
-                subRoller.addContexts(...this.#components);
+                subRoller.addContexts(...this.components);
                 // Roll it
                 await subRoller.roll();
                 // Get sub result
@@ -215,7 +211,7 @@ export class TableRoller extends GenericFileRoller<string> {
                         return "ERROR";
                     }
                     const rollsRoller = roller as StackRoller;
-                    rollsRoller.addContexts(...this.#components);
+                    rollsRoller.addContexts(...this.components);
                     await rollsRoller.roll();
                     this.rolls = rollsRoller.result;
                     if (!rollsRoller.isStatic) {
@@ -360,7 +356,7 @@ export class TableRoller extends GenericFileRoller<string> {
                 );
                 if (maybeRoller.isSome()) {
                     const roller = maybeRoller.unwrap();
-                    roller.addContexts(...this.#components);
+                    roller.addContexts(...this.components);
                     if (roller instanceof StackRoller) {
                         this.lookupRoller = roller;
